% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_showchanges.R
\name{icd_showchanges_icd3}
\alias{icd_showchanges_icd3}
\title{Show all changes in ICD history relating to the
3-digit codes contained in a given vector icd}
\usage{
icd_showchanges_icd3(icd3)
}
\arguments{
\item{icd3}{Vector of three-digit ICD codes}
}
\value{
data.frame with columns YEAR, ICD_CODE, ICD_LABEL and, if specified, DIAG_GROUP
}
\description{
Show all changes in ICD history relating to the
3-digit codes contained in a given vector icd
}
