% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_parse.R
\name{is_icd_code}
\alias{is_icd_code}
\title{Test whether a string is an ICD code}
\usage{
is_icd_code(str)
}
\arguments{
\item{str}{Character vector to be tested}
}
\value{
Logical vector the same length as the character input
}
\description{
An ICD code consists of a three digit code
(i.e. one upper-case letter followed by
two digits), optionally with a two digit
subcode (which may or may not have a period
separating the three-digit code and subcode)
}
