% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_lookup.R
\name{icd_browse}
\alias{icd_browse}
\title{Lookup a 3-digit ICD-10-GM code in the official BfArM browser}
\usage{
icd_browse(icd3, year = NULL)
}
\arguments{
\item{icd3}{3-digit ICD-10-GM code (e.g. "A01")}

\item{year}{ICD-10-GM version (default: most recent available version). Only works for year >= 2009.}
}
\value{
Called for side-effect, but returns the URL invisibly
}
\description{
Given a 3-digit ICD-10-GM code, this function will generate the URL of the
corresponding page of the BfArM ICD-10-GM browser, and use \code{browseURL()} to
open it.
}
\details{
This currently provides the correct URL for ICD-10-GM versions from 2009.
Given that BfArM are still using the old dimdi.de domain for this purpose,
it is possible that the URL schema will change in the near future.

This is a convenience function intended for interactive use.
}
\examples{
icd_browse("E10")
}
\seealso{
\code{\link[=icd_lookup]{icd_lookup()}} to lookup one or more codes in the R console
}
