% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_expand.R
\name{icd_expand}
\alias{icd_expand}
\title{Expand list of ICD codes to include all possible subcodes}
\usage{
icd_expand(icd_in, year, col_icd = "ICD", col_meta = NULL)
}
\arguments{
\item{icd_in}{Data frame defining ICD codes of interest}

\item{year}{ICD 10 version}

\item{col_icd}{Column of icd_in containing ICD codes (Default: ICD)}

\item{col_meta}{(Optional) Columns containing meta information to retain (e.g. Grouper, age or other criteria for later use). If left NULL, only col_icd is retained.}
}
\value{
data.frame with columns YEAR, ICD_CODE, ICD_COMPRESSED, ICD_LABEL and, if specified, columns specified by col_meta
}
\description{
The function `icd_expand` takes a data.frame containing ICD codes 
and optional metadata as input. It returns a data.frame containing 
all ICD codes at or below the specified level of the hierarchy
(e.g. the specification "E11" is expanded to include all three,
four and five-digit codes beginning with E11).
}
