% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_history.R
\name{icd_history}
\alias{icd_history}
\title{Historize a list of ICD codes to cover the specified years}
\usage{
icd_history(icd_expand, years)
}
\arguments{
\item{icd_expand}{A data.frame (e.g. as generated by the function icd_expand)}

\item{years}{Years to historize (e.g. 2005:2014)}
}
\value{
data.frame with columns YEAR, ICD_CODE, ICD_COMPRESSED, ICD_LABEL and, if specified, DIAG_GROUP
}
\description{
The function `icd_history` takes the result of `icd_expand`, 
specified for a particular year, and returns a data.frame 
containing all corresponding codes for the specified years 
(from 2003). To do this, it applies the ICD-10-GM transition 
tables to map codes between successive ICD-10-GM versions. 
Only automatic transitions are followed.
}
