% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_parse.R
\name{icd_parse}
\alias{icd_parse}
\title{Extract all ICD codes from a character vector}
\usage{
icd_parse(str, bind_rows = TRUE)
}
\arguments{
\item{str}{Character vector from which to extract all ICD codes}

\item{bind_rows}{Whether to convert the matrix output of `stirngi::stri_match_all` to a data.frame, with additional `icd_norm` and `icd_sub`}
}
\value{
data.frame (if bind_rows = TRUE) or matrix
}
\description{
An ICD code consists of a three digit code
(i.e. one upper-case letter followed by
two digits), optionally with a two digit
subcode (which may or may not have a period
separating the three-digit code and subcode)
}
\details{
By default, the function returns a data.frame
containing the matched codes and the standardised
three digit code (`icd3`), subcode (`icd_subcode`),
normcode (`icd_norm`) and code without period (`icd_sub`).

If `bind_rows = FALSE`, the list output of
`stringi::stri_match_all_regex` is returned.
This is particularly useful to retrieve the
matches from each element of the `str` vector
separately.
}
